requirements.txt
flask==2.3.3
Pillow==10.0.0
numpy==1.24.3
rembg==2.0.50
werkzeug==2.3.7
gunicorn==21.2.0
scikit-image==0.21.0
opencv-python-headless==4.8.1.78
python-dotenv==1.0.0